pyproject.toml
[tool.poetry]
name = "manage-scripts-for-nas"
version = "0.1.0"
description = ""
authors = ["Lukas Kotatko <lukas.kotatko@heureka.group>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pysmb = "^1.2.9.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pyfakefs = "^5.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
